requirements.txt
fastapi==0.119.0
uvicorn==0.37.0
flask==3.0.3
PyJWT==2.9.0
